Jahnvi_Tower_of_Hanoi_contribution.c
Tower of Hanoi

The objective of the puzzle is to move the entire stack to another rod, obeying the following simple rules−

Only one disk can be moved at a time.

Each move consists of taking the upper disk from one of the stacks and placing it on top of another stack i.e. a disk can only be moved if it is the uppermost disk on a stack.

No disk may be placed on top of a smaller disk.